Compatibility to moodle 4
